closes #26. move IB to connector specific file
