+++ ta_scanner/data/base_connector.py
@@ -1,0 +1,5 @@
+from abc import ABCMeta, abstractmethod
+
+# python3
+class DataFetcherBase(object, metaclass=ABCMeta):
+    pass
