+++ ta_scanner/data/data.py
@@ -3,8 +3,6 @@ import pandas as pd
 import numpy as np
 import os
 from loguru import logger
-from ib_insync import IB, Forex, Future, ContFuture, Stock, Contract
-from ib_insync import util as ib_insync_util
 from psycopg2 import sql
 
 # from datetime import datetime, timedelta, timezone, date
@@ -14,219 +12,14 @@ from trading_calendars import get_calendar, TradingCalendar
 from typing import Optional, Dict, Any, List, Tuple
 
 from ta_scanner.models import gen_engine, init_db, Quote
-
-
-class TimezoneNames(Enum):
-    US_EASTERN = "US/Eastern"
-    US_CENTRAL = "US/Central"
-    US_MOUNTAIN = "US/Mountain"
-    US_PACIFIC = "US/Pacific"
-    UTC = "UTC"
-
-
-class WhatToShow(Enum):
-    TRADES = "TRADES"
-    MIDPOINT = "MIDPOINT"
-    BID = "BID"
-    ASK = "ASK"
-    BID_ASK = "BID_ASK"
-    ADJUSTED_LAST = "ADJUSTED_LAST"
-    HISTORICAL_VOLATILITY = "HISTORICAL_VOLATILITY"
-    OPTION_IMPLIED_VOLATILITY = "OPTION_IMPLIED_VOLATILITY"
-    REBATE_RATE = "REBATE_RATE"
-    FEE_RATE = "FEE_RATE"
-    YIELD_BID = "YIELD_BID"
-    YIELD_ASK = "YIELD_ASK"
-    YIELD_BID_ASK = "YIELD_BID_ASK"
-    YIELD_LAST = "YIELD_LAST"
-
-
-class Exchange(Enum):
-    SMART = "SMART"
-    NYSE = "NYSE"
-    GLOBEX = "GLOBEX"
-    NYMEX = "NYMEX"
-    ECBOT = "ECBOT"
-    CBOE = "CBOE"
-    ICE = "ICE"
-
-
-class Calendar(Enum):
-    # https://github.com/quantopian/trading_calendars
-    DEFAULT = "XNYS"  # default to NYSE
-    NYSE = "XNYS"
-    CME = "CMES"
-    CBOE = "XCBF"
-    ICE = "IEPA"
-
-    @staticmethod
-    def futures_lookup_hash() -> Dict:
-        return {
-            Calendar.CME: [
-                # equities
-                "/ES",
-                "/MES",
-                "/MNQ",
-                "/NQ",
-                "/MNQ",
-                # metals
-                "/GC",
-                "/MGC",
-                # energy
-                "/CL",
-                "/QM",
-                # currencies
-                "/M6A",
-                "/M6B",
-                "/M6E",
-                # interest rates
-                "/GE",
-                "/ZN",
-                "/ZN",
-                "/ZT",
-                # grains
-                "/ZC",
-                "/YC",
-                "/ZS",
-                "/YK",
-                "/ZW",
-                "/YW",
-            ],
-            Calendar.CBOE: [],
-            Calendar.ICE: [],
-        }
-
-    @staticmethod
-    def select_exchange_by_symbol(symbol: str):
-        for k, v in Calendar.futures_lookup_hash().items():
-            if symbol in v:
-                return k
-        logger.warning(f"Did not find a calendar entry for symbol={symbol}")
-        return Calendar.DEFAULT
-
-    @staticmethod
-    def init_by_symbol(symbol: str) -> TradingCalendar:
-        if "/" in symbol:
-            key = Calendar.select_exchange_by_symbol(symbol)
-            name = key.value
-        else:
-            name = Calendar.NYSE.value
-        return get_calendar(name)
-
-
-class Currency(Enum):
-    USD = "USD"
-
-
-from abc import ABCMeta, abstractmethod
-
-# python3
-class DataFetcherBase(object, metaclass=ABCMeta):
-    pass
-
-
-class IbDataFetcher(DataFetcherBase):
-    def __init__(self, client_id: int = 0):
-        self.ib = None
-        self.client_id = client_id
-
-    def _init_client(
-        self, host: str = "127.0.0.1", port: int = 4001
-    ) -> None:
-        ib = IB()
-        ib.connect(host, port, clientId=self.client_id)
-        self.ib = ib
-
-    def _execute_req_historical(
-        self, contract, dt, duration, bar_size_setting, what_to_show, use_rth
-    ) -> pd.DataFrame:
-        if self.ib is None or not self.ib.isConnected():
-            self._init_client()
-
-        dfs = []
-        for rth in [True, False]:
-            bars = self.ib.reqHistoricalData(
-                contract,
-                endDateTime=dt,
-                durationStr=duration,
-                barSizeSetting=bar_size_setting,
-                whatToShow=what_to_show,
-                useRTH=rth,  # use_rth,
-                formatDate=2,  # return as UTC time
-            )
-            x = ib_insync_util.df(bars)
-            x['rth'] = rth
-            dfs.append(x)
-        df = pd.concat(dfs).drop_duplicates().reset_index(drop=True)
-        return df
-
-    def request_stock_instrument(
-        self, instrument_symbol: str, dt: datetime.datetime, what_to_show: str
-    ) -> pd.DataFrame:
-        exchange = Exchange.SMART.value
-        contract = Stock(instrument_symbol, exchange, Currency.USD.value)
-        duration = "2 D"
-        bar_size_setting = "1 min"
-        use_rth = False
-        return self._execute_req_historical(
-            contract, dt, duration, bar_size_setting, what_to_show, use_rth
-        )
-
-    def select_echange_by_symbol(self, symbol):
-        d = {
-            Exchange.GLOBEX: [
-                # equities
-                "/ES",
-                "/MES",
-                "/MNQ",
-                "/NQ",
-                "/MNQ"
-                # currencies
-                "/M6A",
-                "/M6B",
-                "/M6E",
-                # # interest rates
-                # ? '/GE', '/ZN', '/ZN', '/ZT',
-            ],
-            Exchange.ECBOT: ["/ZC", "/YC", "/ZS", "/YK", "/ZW", "/YW"],
-            Exchange.NYMEX: ["/GC", "/MGC", "/CL", "/QM",],
-        }
-
-        for k, v in d.items():
-            if symbol in v:
-                return k
-        raise NotImplementedError
-
-    def request_future_instrument(
-        self,
-        symbol: str,
-        dt: datetime.datetime,
-        what_to_show: str,
-        contract_date: Optional[str] = None,
-    ) -> pd.DataFrame:
-        exchange_name = self.select_echange_by_symbol(symbol).value
-
-        if contract_date:
-            raise NotImplementedError
-        else:
-            contract = ContFuture(symbol, exchange_name, currency=Currency.USD.value)
-
-        duration = "1 D"
-        bar_size_setting = "1 min"
-        use_rth = False
-        return self._execute_req_historical(
-            contract, dt, duration, bar_size_setting, what_to_show, use_rth
-        )
-
-    def request_instrument(
-        self, symbol: str, dt, what_to_show, contract_date: Optional[str] = None,
-    ):
-        if "/" in symbol:
-            return self.request_future_instrument(
-                symbol, dt, what_to_show, contract_date
-            )
-        else:
-            return self.request_stock_instrument(symbol, dt, what_to_show)
+from ta_scanner.data.base_connector import DataFetcherBase
+from ta_scanner.data.constants import (
+    TimezoneNames,
+    WhatToShow,
+    Exchange,
+    Calendar,
+    Currency,
+)
 
 
 def extract_kwarg(kwargs: Dict, key: str, default_value: Any = None) -> Optional[Any]:
